Rename con.js to connection.js
rename con to connection
